feat(vm): W8a — positions are objects, not indices (§6.2)
CR 6.2.6 is explicit that the Runner's current position "is a specific
element of the sequence of positions, not an index into that sequence".
The kernel held an index, so every change to the ice protecting the
attacked server silently moved the Runner. Positions are now objects:

- `object::IcePosition { id, ice }`; `CoreState.ice` is each server's
  SEQUENCE OF POSITIONS, innermost first, and `RunCtx.position` is a
  position id.
- 6.2.2 creation with its four placements (a outermost, b innermost,
  c directly inward, d/e not yet expressible) — and 6.2.2f, where a swap
  creates none, the two ice simply re-occupying the existing positions.
  For an install the position is created at step 8.5.16b and occupied at
  8.5.16e, exactly as the rule says.
- 6.2.4 destruction is a REAL checkpoint step 10.3.1i, with both of its
  exceptions: the Runner's current position, and a server with an
  installation in progress protecting it.
- 6.2.3 "same position" as `TargetFilter::IceInSamePositionAs(ref)` —
  the rule concept is the variant, the reference position (source, or
  the Runner) is the content.
- 6.2.7a/c/d as `Vm::apply_ice_change_to_run`, run at the top of the
  checkpoint; 6.2.7e needs no code at all, which is the whole point.
- `Instruction::MoveIce` (6.2.2, ice moved to a position) and
  `Instruction::MoveRunnerToIce` (6.2.8a-d, including the 6.2.5d/6.2.8c
  no-op during the Success and Run Ends phases).

Two defects fell out: `swap_cards` no-opped on two pieces of ice
protecting the SAME server (same zone, different positions), and
`announcement_for` filtered candidates with no source, so no
source-relative criterion could ever be announced against.

DP-7a 145 -> 152, DP-7b 472 -> 489.

+++ crates/jinteki-cr/src/frames.rs
@@ -28,10 +28,11 @@ pub struct RunCtx {
     pub run_id: u64,
     /// CR 6.1.2: the attacked server (announced at initiation).
     pub server: ServerId,
-    /// CR 6.2: the Runner's position, as an index into the server's ice
-    /// list counted from the innermost (1-based). `None` = no position /
-    /// inside the server.
-    pub position: Option<usize>,
+    /// CR 6.2.6: the Runner's position — the IDENTITY of one element of the
+    /// attacked server's sequence of positions, never an index into it, so
+    /// adding or removing other positions cannot move the Runner. `None` =
+    /// no position (6.2.5c/d/e).
+    pub position: Option<u64>,
     /// Whether the current approach/encounter arrived from phase 2/3 (drives
     /// `step_pass_ice`'s "if the run got here from (2) or (3)").
     pub came_from_ice: bool,
